requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2
tensorflow
xgboost
python-dotenv
scikit-learn
